refactor(prompt): extract advanced studio slice
