+++ nodes.py
@@ -8,8 +8,8 @@ import os
 import random
 import re
 import sys
-from math import ceil, isfinite, sqrt
-from typing import Any, Optional
+from math import ceil, sqrt
+from typing import Any
 
 try:
     from .easyuse_anima.common.serialization import (
@@ -61,6 +61,42 @@ try:
         _prompt_data_output as _prompt_data_output,
         _prompt_data_parameter_snapshot as _prompt_data_parameter_snapshot,
         _set_prompt_data_output as _set_prompt_data_output,
+    )
+    from .easyuse_anima.prompt.advanced import (
+        ADVANCED_FIELDS_WORKFLOW_PROPERTY as ADVANCED_FIELDS_WORKFLOW_PROPERTY,
+        ADVANCED_FIELD_LABELS as ADVANCED_FIELD_LABELS,
+        ADVANCED_FIELD_PANES as ADVANCED_FIELD_PANES,
+        ADVANCED_FIELD_TYPES as ADVANCED_FIELD_TYPES,
+        EXTEND_PROMPT_SLOT_SPECS as EXTEND_PROMPT_SLOT_SPECS,
+        PROMPT_STUDIO_LEGACY_FIXED_WILDCARD_MODES as PROMPT_STUDIO_LEGACY_FIXED_WILDCARD_MODES,
+        PROMPT_STUDIO_ADVANCED_RETURN_NAMES as PROMPT_STUDIO_ADVANCED_RETURN_NAMES,
+        PROMPT_STUDIO_ADVANCED_RETURN_TYPES as PROMPT_STUDIO_ADVANCED_RETURN_TYPES,
+        PROMPT_STUDIO_WILDCARD_SEED_CONTROL_ALIASES as PROMPT_STUDIO_WILDCARD_SEED_CONTROL_ALIASES,
+        _advanced_artist_field_prompt as _advanced_artist_field_prompt,
+        _advanced_default_fields as _advanced_default_fields,
+        _advanced_enabled_naia_panes as _advanced_enabled_naia_panes,
+        _advanced_enabled_pane_fields as _advanced_enabled_pane_fields,
+        _advanced_field_input_values as _advanced_field_input_values,
+        _advanced_field_socket_name as _advanced_field_socket_name,
+        _advanced_fields_json as _advanced_fields_json,
+        _advanced_fields_with_artist_override as _advanced_fields_with_artist_override,
+        _advanced_has_enabled_naia as _advanced_has_enabled_naia,
+        _advanced_pane_parts as _advanced_pane_parts,
+        _advanced_prompt_data_fields as _advanced_prompt_data_fields,
+        _advanced_prompt_with_artist_override as _advanced_prompt_with_artist_override,
+        _advanced_uses_naia_resolution as _advanced_uses_naia_resolution,
+        _apply_advanced_field_inputs as _apply_advanced_field_inputs,
+        _as_advanced_height as _as_advanced_height,
+        _bind_advanced_runtime as _bind_advanced_runtime,
+        _build_advanced_prompt_data as _build_advanced_prompt_data,
+        _build_advanced_prompts as _build_advanced_prompts,
+        _clone_advanced_fields as _clone_advanced_fields,
+        _correct_advanced_field_sequence as _correct_advanced_field_sequence,
+        _expand_advanced_wildcard_fields as _expand_advanced_wildcard_fields,
+        _normalize_advanced_fields as _normalize_advanced_fields,
+        _normalize_prompt_studio_wildcard_seed_control as _normalize_prompt_studio_wildcard_seed_control,
+        _set_naia_field_text as _set_naia_field_text,
+        _translate_prompt_fields as _translate_prompt_fields,
     )
     from .easyuse_anima.prompt.regional import (
         REGIONAL_CONFIG_VERSION as REGIONAL_CONFIG_VERSION,
@@ -191,6 +227,12 @@ try:
         EasyUseAnimaPromptDataConditioning as EasyUseAnimaPromptDataConditioning,
         EasyUseAnimaPromptDataUnpack as EasyUseAnimaPromptDataUnpack,
         _bind_prompt_data_node_runtime as _bind_prompt_data_node_runtime,
+    )
+    from .easyuse_anima.nodes.prompt_advanced_nodes import (
+        EasyUseAnimaPromptStudioAdvanced as EasyUseAnimaPromptStudioAdvanced,
+        EasyUseAnimaPromptStudioAdvancedV2 as EasyUseAnimaPromptStudioAdvancedV2,
+        EasyUseAnimaPromptStudioExtend as EasyUseAnimaPromptStudioExtend,
+        _bind_prompt_advanced_node_runtime as _bind_prompt_advanced_node_runtime,
     )
     from .easyuse_anima.image.geometry import (
         _align_down as _align_down,
@@ -409,6 +451,42 @@ except ImportError:  # allows simple local import tests outside ComfyUI's packag
         _prompt_data_parameter_snapshot as _prompt_data_parameter_snapshot,
         _set_prompt_data_output as _set_prompt_data_output,
     )
+    from easyuse_anima.prompt.advanced import (
+        ADVANCED_FIELDS_WORKFLOW_PROPERTY as ADVANCED_FIELDS_WORKFLOW_PROPERTY,
+        ADVANCED_FIELD_LABELS as ADVANCED_FIELD_LABELS,
+        ADVANCED_FIELD_PANES as ADVANCED_FIELD_PANES,
+        ADVANCED_FIELD_TYPES as ADVANCED_FIELD_TYPES,
+        EXTEND_PROMPT_SLOT_SPECS as EXTEND_PROMPT_SLOT_SPECS,
+        PROMPT_STUDIO_LEGACY_FIXED_WILDCARD_MODES as PROMPT_STUDIO_LEGACY_FIXED_WILDCARD_MODES,
+        PROMPT_STUDIO_ADVANCED_RETURN_NAMES as PROMPT_STUDIO_ADVANCED_RETURN_NAMES,
+        PROMPT_STUDIO_ADVANCED_RETURN_TYPES as PROMPT_STUDIO_ADVANCED_RETURN_TYPES,
+        PROMPT_STUDIO_WILDCARD_SEED_CONTROL_ALIASES as PROMPT_STUDIO_WILDCARD_SEED_CONTROL_ALIASES,
+        _advanced_artist_field_prompt as _advanced_artist_field_prompt,
+        _advanced_default_fields as _advanced_default_fields,
+        _advanced_enabled_naia_panes as _advanced_enabled_naia_panes,
+        _advanced_enabled_pane_fields as _advanced_enabled_pane_fields,
+        _advanced_field_input_values as _advanced_field_input_values,
+        _advanced_field_socket_name as _advanced_field_socket_name,
+        _advanced_fields_json as _advanced_fields_json,
+        _advanced_fields_with_artist_override as _advanced_fields_with_artist_override,
+        _advanced_has_enabled_naia as _advanced_has_enabled_naia,
+        _advanced_pane_parts as _advanced_pane_parts,
+        _advanced_prompt_data_fields as _advanced_prompt_data_fields,
+        _advanced_prompt_with_artist_override as _advanced_prompt_with_artist_override,
+        _advanced_uses_naia_resolution as _advanced_uses_naia_resolution,
+        _apply_advanced_field_inputs as _apply_advanced_field_inputs,
+        _as_advanced_height as _as_advanced_height,
+        _bind_advanced_runtime as _bind_advanced_runtime,
+        _build_advanced_prompt_data as _build_advanced_prompt_data,
+        _build_advanced_prompts as _build_advanced_prompts,
+        _clone_advanced_fields as _clone_advanced_fields,
+        _correct_advanced_field_sequence as _correct_advanced_field_sequence,
+        _expand_advanced_wildcard_fields as _expand_advanced_wildcard_fields,
+        _normalize_advanced_fields as _normalize_advanced_fields,
+        _normalize_prompt_studio_wildcard_seed_control as _normalize_prompt_studio_wildcard_seed_control,
+        _set_naia_field_text as _set_naia_field_text,
+        _translate_prompt_fields as _translate_prompt_fields,
+    )
     from easyuse_anima.prompt.regional import (
         REGIONAL_CONFIG_VERSION as REGIONAL_CONFIG_VERSION,
         REGIONAL_CONFIG_WORKFLOW_PROPERTY as REGIONAL_CONFIG_WORKFLOW_PROPERTY,
@@ -538,6 +616,12 @@ except ImportError:  # allows simple local import tests outside ComfyUI's packag
         EasyUseAnimaPromptDataConditioning as EasyUseAnimaPromptDataConditioning,
         EasyUseAnimaPromptDataUnpack as EasyUseAnimaPromptDataUnpack,
         _bind_prompt_data_node_runtime as _bind_prompt_data_node_runtime,
+    )
+    from easyuse_anima.nodes.prompt_advanced_nodes import (
+        EasyUseAnimaPromptStudioAdvanced as EasyUseAnimaPromptStudioAdvanced,
+        EasyUseAnimaPromptStudioAdvancedV2 as EasyUseAnimaPromptStudioAdvancedV2,
+        EasyUseAnimaPromptStudioExtend as EasyUseAnimaPromptStudioExtend,
+        _bind_prompt_advanced_node_runtime as _bind_prompt_advanced_node_runtime,
     )
     from easyuse_anima.image.geometry import (
         _align_down as _align_down,
@@ -708,34 +792,8 @@ except ImportError:  # allows simple local import tests outside ComfyUI's packag
 
 logger = logging.getLogger("ComfyUI-EasyUseAnima")
 
-ADVANCED_FIELD_TYPES = {"quality", "artist", "trigger", "general", "naia"}
-ADVANCED_FIELD_PANES = {"positive", "negative"}
-ADVANCED_FIELD_LABELS = {
-    "quality": "Quality Tags",
-    "artist": "Artist Tags",
-    "trigger": "Trigger Words",
-    "general": "General Tags",
-    "naia": "NAIA Prompt",
-}
-ADVANCED_FIELDS_WORKFLOW_PROPERTY = "easyuse_anima_advanced_fields"
 WILDCARD_RESERVED_NEXT_SEED_INPUT = "easyuse_anima_reserved_wildcard_next_seed"
 WILDCARD_QUEUE_MAX_SAFE_SEED = PUBLIC_MAX_SEED
-PROMPT_STUDIO_WILDCARD_SEED_CONTROL_ALIASES = {
-    "fixed": SEED_CONTROL_FIXED,
-    "고정": SEED_CONTROL_FIXED,
-    "random": SEED_CONTROL_RANDOMIZE,
-    "randomize": SEED_CONTROL_RANDOMIZE,
-    "매번 랜덤": SEED_CONTROL_RANDOMIZE,
-    "increase": SEED_CONTROL_INCREMENT,
-    "increment": SEED_CONTROL_INCREMENT,
-    "증가": SEED_CONTROL_INCREMENT,
-}
-PROMPT_STUDIO_LEGACY_FIXED_WILDCARD_MODES = {
-    WILDCARD_MODE_FIXED,
-    "고정",
-    "reproduce",
-    "재현",
-}
 EASY_USE_ANIMA_INPUT_TYPE = "EASY_USE_ANIMA_INPUT"
 EASY_USE_ANIMA_INPUT_SCHEMA = "easy_use_anima_input"
 EASY_USE_ANIMA_INPUT_SETTINGS_VERSION = 1
@@ -1220,23 +1278,6 @@ AIO_GENERATION_DEFAULT_SETTINGS = {
         "feed_count": 12,
     },
 }
-EXTEND_PROMPT_SLOT_SPECS = [
-    ("quality_tags_1", "positive", "quality", "Quality Tags 1", DEFAULT_QUALITY_TAGS, 72),
-    ("quality_tags_2", "positive", "quality", "Quality Tags 2", "", 72),
-    ("naia_prompt_3", "positive", "general", "NAIA Prompt 3", "", 150),
-    ("general_tags_4", "positive", "general", "General Tags 4", "", 120),
-    ("general_tags_5", "positive", "general", "General Tags 5", "", 120),
-    ("general_tags_6", "positive", "general", "General Tags 6", "", 120),
-    ("general_tags_7", "positive", "general", "General Tags 7", "", 120),
-    ("general_tags_8", "positive", "general", "General Tags 8", "", 120),
-    ("general_tags_9", "positive", "general", "General Tags 9", "", 120),
-    ("trailing_tags_10", "positive", "general", "Trailing Quality Tags 10", DEFAULT_TRAILING_QUALITY_TAGS, 72),
-    ("trailing_tags_11", "positive", "general", "Trailing Quality Tags 11", "", 72),
-    ("negative_prompt_1", "negative", "quality", "Negative Prompt 1", "", 120),
-    ("negative_prompt_2", "negative", "quality", "Negative Prompt 2", "", 120),
-    ("negative_prompt_3", "negative", "general", "Negative Prompt 3", "", 120),
-    ("negative_prompt_4", "negative", "general", "Negative Prompt 4", "", 120),
-]
 AIO_FINAL_UPSCALE_BACKENDS = ("usdu", "resshift")
 AIO_USDU_MODE_TYPES = ("Linear", "Chess", "None")
 AIO_USDU_SEAM_FIX_MODES = ("None", "Band Pass", "Half Tile", "Half Tile + Intersections")
@@ -1249,10 +1290,6 @@ AIO_RESHIFT_DTYPES = ("bf16", "fp32")
 
 
 _TRIGGER_WORD_KEYS = ("trainedWords", "trained_words", "trigger_words", "activation_text")
-_ADVANCED_FIELD_SOCKET_PREFIX = "field_"
-_ADVANCED_FIELD_SOCKET_RE = re.compile(r"[^A-Za-z0-9_]")
-
-
 class _AnyType(str):
     def __ne__(self, __value: object) -> bool:
         return False
@@ -4451,685 +4488,6 @@ def _call_impact_detailer(detailer, **kwargs):
     return method(**call_kwargs)
 
 
-def _translate_prompt_fields(fields: list[dict]) -> list[dict]:
-    translated: list[dict] = []
-    for field in fields:
-        item = dict(field)
-        text = str(item.get("text") or "")
-        if text and has_prompt_translation_markers(text):
-            item["text"] = _translate_prompt_text(text)
-        translated.append(item)
-    return translated
-
-
-def _advanced_default_fields() -> list[dict]:
-    return [
-        {
-            "id": "positive_quality",
-            "pane": "positive",
-            "type": "quality",
-            "label": ADVANCED_FIELD_LABELS["quality"],
-            "text": DEFAULT_QUALITY_TAGS,
-            "height": 72,
-            "enabled": True,
-        },
-        {
-            "id": "positive_artist",
-            "pane": "positive",
-            "type": "artist",
-            "label": ADVANCED_FIELD_LABELS["artist"],
-            "text": "",
-            "height": 72,
-            "enabled": True,
-        },
-        {
-            "id": "positive_trigger",
-            "pane": "positive",
-            "type": "trigger",
-            "label": ADVANCED_FIELD_LABELS["trigger"],
-            "text": "",
-            "height": 72,
-            "enabled": True,
-            "pin": True,
-        },
-        {
-            "id": "positive_general",
-            "pane": "positive",
-            "type": "general",
-            "label": ADVANCED_FIELD_LABELS["general"],
-            "text": "",
-            "height": 150,
-            "enabled": True,
-        },
-        {
-            "id": "positive_trailing",
-            "pane": "positive",
-            "type": "general",
-            "label": ADVANCED_FIELD_LABELS["general"],
-            "text": DEFAULT_TRAILING_QUALITY_TAGS,
-            "height": 72,
-            "enabled": True,
-        },
-        {
-            "id": "negative_general",
-            "pane": "negative",
-            "type": "general",
-            "label": ADVANCED_FIELD_LABELS["general"],
-            "text": "",
-            "height": 120,
-            "enabled": True,
-        },
-    ]
-
-
-def _advanced_fields_json(fields: list[dict] | None = None) -> str:
-    return json.dumps(
-        fields if fields is not None else _advanced_default_fields(),
-        ensure_ascii=False,
-        separators=(",", ":"),
-    )
-
-
-def _as_advanced_height(value, default: int = 72) -> int:
-    return max(36, _as_int(value, default))
-
-
-def _normalize_advanced_fields(value: str | list | None) -> list[dict]:
-    raw = value
-    if isinstance(value, str):
-        try:
-            raw = json.loads(value or "[]")
-        except json.JSONDecodeError:
-            raw = []
-    if not isinstance(raw, list):
-        raw = []
-    if not raw:
-        raw = _advanced_default_fields()
-
-    fields: list[dict] = []
-    seen_naia_panes: set[str] = set()
-    seen_trigger = False
-    for index, item in enumerate(raw):
-        if not isinstance(item, dict):
-            continue
-        pane = str(item.get("pane") or "positive").strip().lower()
-        if pane not in ADVANCED_FIELD_PANES:
-            pane = "positive"
-        field_type = str(item.get("type") or "general").strip().lower()
-        if field_type not in ADVANCED_FIELD_TYPES:
-            field_type = "general"
-        if pane == "negative" and field_type == "trigger":
-            field_type = "general"
-        if field_type == "naia":
-            if pane in seen_naia_panes:
-                continue
-            seen_naia_panes.add(pane)
-        if field_type == "trigger":
-            if seen_trigger:
-                continue
-            seen_trigger = True
-            pane = "positive"
-        default_label = ADVANCED_FIELD_LABELS.get(field_type, ADVANCED_FIELD_LABELS["general"])
-        label = str(item.get("label") or default_label).strip() or default_label
-        field_id = str(item.get("id") or f"{pane}_{field_type}_{index + 1}").strip()
-        if not field_id:
-            field_id = f"{pane}_{field_type}_{index + 1}"
-        fields.append({
-            "id": field_id,
-            "pane": pane,
-            "type": field_type,
-            "label": label,
-            "text": str(item.get("text") or ""),
-            "height": _as_advanced_height(item.get("height"), 72),
-            "enabled": _as_bool(item.get("enabled"), True),
-            "pin": _as_bool(item.get("pin"), field_type == "trigger"),
-        })
-
-    return fields or _advanced_default_fields()
-
-
-def _clone_advanced_fields(fields: list[dict]) -> list[dict]:
-    return [dict(field) for field in fields]
-
-
-def _advanced_field_socket_name(field: dict) -> str:
-    raw = _ADVANCED_FIELD_SOCKET_RE.sub("_", str(field.get("id") or "field")).strip("_")
-    return f"{_ADVANCED_FIELD_SOCKET_PREFIX}{raw or 'field'}"
-
-
-def _advanced_field_input_values(field_inputs: dict) -> dict[str, str]:
-    values: dict[str, str] = {}
-    for key, value in field_inputs.items():
-        if not str(key).startswith(_ADVANCED_FIELD_SOCKET_PREFIX):
-            continue
-        single = _single_value(value)
-        if single is None:
-            continue
-        values[str(key)] = str(single)
-    return values
-
-
-def _apply_advanced_field_inputs(fields: list[dict], field_inputs: dict) -> list[dict]:
-    values = _advanced_field_input_values(field_inputs)
-    if not values:
-        return _clone_advanced_fields(fields)
-
-    effective = _clone_advanced_fields(fields)
-    for field in effective:
-        name = _advanced_field_socket_name(field)
-        if name in values:
-            field["text"] = values[name]
-    return effective
-
-
-def _advanced_enabled_naia_panes(fields: list[dict]) -> set[str]:
-    return {
-        str(field.get("pane") or "positive")
-        for field in fields
-        if field.get("type") == "naia" and field.get("enabled") is not False
-    }
-
-
-def _advanced_has_enabled_naia(fields: list[dict]) -> bool:
-    return bool(_advanced_enabled_naia_panes(fields))
-
-
-def _advanced_uses_naia_resolution(bucket) -> bool:
-    return _normalize_resolution_bucket(bucket) == NAIA_ADVANCED_RESOLUTION_BUCKET
-
-
-def _set_naia_field_text(fields: list[dict], pane: str, prompt: str) -> list[dict]:
-    normalized = _normalize_advanced_fields(fields)
-    for field in normalized:
-        if field["pane"] == pane and field["type"] == "naia":
-            field["text"] = prompt
-            field["enabled"] = True
-            return normalized
-    return normalized
-
-
-def _advanced_pane_parts(fields: list[dict], pane: str) -> dict[str, list[str]]:
-    parts = {"quality": [], "artist": [], "trigger_fixed": [], "trigger_auto": [], "body": []}
-    for field in fields:
-        if not _as_bool(field.get("enabled"), True):
-            continue
-        if field.get("pane") != pane:
-            continue
-        field_type = field.get("type")
-        text = str(field.get("text") or "")
-        if field_type == "quality":
-            parts["quality"].append(text)
-        elif field_type == "artist":
-            parts["artist"].append(_artist_mix_inline_prompt(text))
-        elif field_type == "trigger":
-            if _as_bool(field.get("pin"), True):
-                parts["trigger_fixed"].append(text)
-            else:
-                parts["trigger_auto"].append(text)
-        else:
-            parts["body"].append(text)
-    return parts
-
-
-def _advanced_enabled_pane_fields(fields: list[dict], pane: str) -> list[dict]:
-    return [
-        field
-        for field in fields
-        if _as_bool(field.get("enabled"), True) and field.get("pane") == pane
-    ]
-
-
-def _correct_advanced_field_sequence(
-    fields: list[dict],
-    include_quality: bool,
-    artist_overrides: str,
-    force_pin_triggers: bool = False,
-) -> str:
-    chunks: list[str] = []
-    pending: list[str] = []
-
-    def flush_pending() -> None:
-        if not pending:
-            return
-        corrected = _correct_builder_prompt(
-            _join_prompt_tokens(*pending),
-            artist_overrides=artist_overrides,
-        )
-        if corrected:
-            chunks.append(corrected)
-        pending.clear()
-
-    for field in fields:
-        field_type = field.get("type")
-        text = str(field.get("text") or "")
-        if field_type == "quality" and not include_quality:
-            continue
-        if field_type == "artist":
-            text = _artist_mix_inline_prompt(text)
-        if field_type == "trigger" and (
-            _as_bool(field.get("pin"), True) or force_pin_triggers
-        ):
-            flush_pending()
-            trigger_prompt = _join_prompt_tokens(text)
-            if trigger_prompt:
-                chunks.append(trigger_prompt)
-            continue
-        pending.append(text)
-
-    flush_pending()
-    return _join_prompt_tokens(*chunks)
-
-
-def _build_advanced_prompts(
-    fields: list[dict],
-    use_anima_mod_guidance: bool,
-    use_negative_anima_mod_guidance: bool,
-    pin_trigger_tags_to_front: bool,
-) -> tuple[str, str, str, str, bool, bool, str, str]:
-    use_amg = _as_bool(use_anima_mod_guidance, False)
-    use_negative_amg = _as_bool(use_negative_anima_mod_guidance, False)
-    force_pin_triggers = _as_bool(pin_trigger_tags_to_front, False)
-    positive = _advanced_pane_parts(fields, "positive")
-    negative = _advanced_pane_parts(fields, "negative")
-    positive_fields = _advanced_enabled_pane_fields(fields, "positive")
-    negative_fields = _advanced_enabled_pane_fields(fields, "negative")
-
-    quality_prompt = _join_prompt_tokens(*positive["quality"])
-    artist_prompt = _join_prompt_tokens(*positive["artist"])
-    regular_prompt = _correct_advanced_field_sequence(
-        positive_fields,
-        include_quality=True,
-        artist_overrides=artist_prompt,
-        force_pin_triggers=force_pin_triggers,
-    )
-    amg_prompt = _correct_advanced_field_sequence(
-        positive_fields,
-        include_quality=False,
-        artist_overrides=artist_prompt,
-        force_pin_triggers=force_pin_triggers,
-    )
-    metadata_prompt = regular_prompt
-
-    negative_quality_prompt = _join_prompt_tokens(*negative["quality"])
-    negative_artist_prompt = _join_prompt_tokens(*negative["artist"])
-    negative_regular_prompt = _correct_advanced_field_sequence(
-        negative_fields,
-        include_quality=True,
-        artist_overrides=negative_artist_prompt,
-    )
-    negative_amg_prompt = _correct_advanced_field_sequence(
-        negative_fields,
-        include_quality=False,
-        artist_overrides=negative_artist_prompt,
-    )
-
-    filter_words = resolve_metadata_filter_words()
-    metadata_prompt = _filter_metadata_prompt(metadata_prompt, filter_words)
-    metadata_negative_prompt = _filter_metadata_prompt(negative_regular_prompt, filter_words)
-    output_prompt = amg_prompt if use_amg else regular_prompt
-    output_negative_prompt = negative_amg_prompt if use_negative_amg else negative_regular_prompt
-    return (
-        output_prompt,
-        output_negative_prompt,
-        quality_prompt,
-        negative_quality_prompt,
-        use_amg,
-        use_negative_amg,
-        metadata_prompt,
-        metadata_negative_prompt,
-    )
-
-
-def _expand_advanced_wildcard_fields(
-    fields: list[dict],
-    seed: int,
-    mode: str,
-) -> tuple[list[dict], dict[str, Any]]:
-    mode_key = normalize_prompt_studio_wildcard_mode(mode)
-    expanded_fields = _clone_advanced_fields(fields)
-
-    wildcard_fields = []
-    wildcard_texts = []
-    for field in expanded_fields:
-        text = str(field.get("text") or "")
-        if has_wildcard_syntax(text):
-            wildcard_fields.append(field)
-            wildcard_texts.append(text)
-
-    expansions = expand_wildcard_texts(
-        wildcard_texts,
-        seed=seed,
-        mode=mode_key,
-    )
-    changed = False
-    used_keys: list[str] = []
-    missing_keys: list[str] = []
-    for field, text, result in zip(
-        wildcard_fields,
-        wildcard_texts,
-        expansions,
-        strict=True,
-    ):
-        if result.text != text:
-            field["text"] = result.text
-            changed = True
-        for key in result.used_keys:
-            if key not in used_keys:
-                used_keys.append(key)
-        for key in result.missing_keys:
-            if key not in missing_keys:
-                missing_keys.append(key)
-
-    return expanded_fields, {
-        "changed": changed,
-        "used_keys": tuple(used_keys),
-        "missing_keys": tuple(missing_keys),
-    }
-
-
-def _advanced_prompt_data_fields(fields: list[dict]) -> list[dict[str, Any]]:
-    output = []
-    for field in fields:
-        output.append({
-            "id": str(field.get("id") or ""),
-            "pane": str(field.get("pane") or "positive"),
-            "type": str(field.get("type") or "general"),
-            "label": str(field.get("label") or ""),
-            "text": str(field.get("text") or ""),
-            "height": _as_advanced_height(field.get("height"), 72),
-            "enabled": _as_bool(field.get("enabled"), True),
-            "pin": _as_bool(field.get("pin"), field.get("type") == "trigger"),
-        })
-    return output
-
-
-def _advanced_artist_field_prompt(fields: list[dict], pane: str) -> str:
-    # Artist data is sourced only from Advanced artist fields, not from @ tags in other fields.
-    return _join_artist_mix_source_prompts(
-        *(
-            str(field.get("text") or "")
-            for field in fields
-            if field.get("pane") == pane
-            and field.get("type") == "artist"
-            and _as_bool(field.get("enabled"), True)
-        )
-    )
-
-
-def _advanced_fields_with_artist_override(fields: list[dict], artist_prompt: str) -> list[dict]:
-    artist_text = _join_prompt_tokens(artist_prompt)
-    output: list[dict] = []
-    inserted = False
-    for field in fields:
-        if field.get("type") == "artist":
-            if artist_text and not inserted:
-                item = dict(field)
-                item["text"] = artist_text
-                output.append(item)
-                inserted = True
-            continue
-        output.append(dict(field))
-
-    if artist_text and not inserted:
-        insert_at = 0
-        for index, field in enumerate(output):
-            if field.get("type") == "quality":
-                insert_at = index + 1
-        output.insert(insert_at, {
-            "id": "artist_mix_override",
-            "pane": "positive",
-            "type": "artist",
-            "label": ADVANCED_FIELD_LABELS["artist"],
-            "text": artist_text,
-            "height": 72,
-            "enabled": True,
-            "pin": False,
-        })
-    return output
-
-
-def _advanced_prompt_with_artist_override(
-    fields: list[dict],
-    artist_prompt: str,
-    include_quality: bool,
-    force_pin_triggers: bool = False,
-) -> str:
-    return _correct_advanced_field_sequence(
-        _advanced_fields_with_artist_override(fields, artist_prompt),
-        include_quality=include_quality,
-        artist_overrides=artist_prompt,
-        force_pin_triggers=force_pin_triggers,
-    )
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-def _build_advanced_prompt_data(
-    compat_result: tuple,
-    effective_fields: list[dict],
-    saved_fields: list[dict],
-    field_inputs: dict[str, str],
-    resolution_bucket: str,
-    resolution_size: str,
-    resolution_custom_width: int,
-    resolution_custom_height: int,
-    wildcard_mode: str,
-    wildcard_seed: int,
-    wildcard_seed_after_generate: str,
-    wildcard_updates: dict[str, Any] | None = None,
-    pin_trigger_tags_to_front: bool = False,
-    parameters: dict[str, Any] | None = None,
-    artist_mix_mode: str = ARTIST_MIX_MODE_OFF,
-    artist_mix_start_percent: float = ARTIST_MIX_DEFAULT_START_PERCENT,
-    artist_mix_strength_scale: float = ARTIST_MIX_DEFAULT_STRENGTH_SCALE,
-    artist_mix_style_gain: float = ARTIST_MIX_DEFAULT_STYLE_GAIN,
-    artist_mix_rms_scale_cap: float = ARTIST_MIX_DEFAULT_RMS_SCALE_CAP,
-    artist_mix_exact_top_k: int = ARTIST_MIX_DEFAULT_EXACT_TOP_K,
-    artist_mix_cluster_count: int = ARTIST_MIX_DEFAULT_CLUSTER_COUNT,
-    artist_mix_dominant_isolation: bool = ARTIST_MIX_DEFAULT_DOMINANT_ISOLATION,
-    artist_mix_dominant_threshold: float = ARTIST_MIX_DEFAULT_DOMINANT_THRESHOLD,
-) -> dict[str, Any]:
-    (
-        positive_prompt,
-        negative_prompt,
-        quality_tags,
-        negative_quality_tags,
-        use_anima_mod_guidance,
-        use_negative_anima_mod_guidance,
-        metadata_prompt,
-        metadata_negative_prompt,
-        width,
-        height,
-    ) = compat_result
-    outputs = {
-        name: value
-        for name, value in zip(EasyUseAnimaPromptStudioAdvanced.RETURN_NAMES, compat_result)
-    }
-    positive_fields = _advanced_enabled_pane_fields(effective_fields, "positive")
-    negative_fields = _advanced_enabled_pane_fields(effective_fields, "negative")
-    positive_artist_prompt = _advanced_artist_field_prompt(effective_fields, "positive")
-    negative_artist_prompt = _advanced_artist_field_prompt(effective_fields, "negative")
-    positive_artist_inline_prompt = _artist_mix_inline_prompt(positive_artist_prompt)
-    negative_artist_inline_prompt = _artist_mix_inline_prompt(negative_artist_prompt)
-    force_pin_triggers = _as_bool(pin_trigger_tags_to_front, False)
-    positive_without_artist = _advanced_prompt_with_artist_override(
-        positive_fields,
-        "",
-        include_quality=not bool(use_anima_mod_guidance),
-        force_pin_triggers=force_pin_triggers,
-    )
-    metadata_prompt_without_artist = _filter_metadata_prompt(
-        _advanced_prompt_with_artist_override(
-            positive_fields,
-            "",
-            include_quality=True,
-            force_pin_triggers=force_pin_triggers,
-        ),
-        resolve_metadata_filter_words(),
-    )
-    negative_without_artist = _advanced_prompt_with_artist_override(
-        negative_fields,
-        "",
-        include_quality=not bool(use_negative_anima_mod_guidance),
-    )
-    selected_artist_mix_mode = _normalize_artist_mix_mode(artist_mix_mode, ARTIST_MIX_MODE_OFF)
-    artist_mix_enabled = selected_artist_mix_mode not in {ARTIST_MIX_MODE_OFF, ARTIST_MIX_MODE_PROMPT}
-    prompt_data_artist_mix_mode = (
-        ARTIST_MIX_MODE_PROMPT
-        if selected_artist_mix_mode == ARTIST_MIX_MODE_OFF
-        else selected_artist_mix_mode
-    )
-    prompt_data_positive_prompt = positive_without_artist if artist_mix_enabled else positive_prompt
-    outputs["positive_prompt"] = prompt_data_positive_prompt
-    wildcard_updates = wildcard_updates or {}
-    parameters = parameters or {}
-    return {
-        "schema": PROMPT_DATA_SCHEMA,
-        "version": PROMPT_DATA_VERSION,
-        "type": PROMPT_DATA_TYPE,
-        "source": "EasyUseAnimaPromptStudioAdvancedV2",
-        "parameters": dict(parameters),
-        "prompt": prompt_data_positive_prompt,
-        "positive_prompt": prompt_data_positive_prompt,
-        "global_prompt": positive_without_artist,
-        "positive_without_artist_section": positive_without_artist,
-        "negative_prompt": negative_prompt,
-        "negative_without_artist_section": negative_without_artist,
-        "metadata_prompt": metadata_prompt,
-        "metadata_prompt_without_artist": metadata_prompt_without_artist,
-        "metadata_negative_prompt": metadata_negative_prompt,
-        "width": int(width),
-        "height": int(height),
-        "pin_trigger_tags_to_front": force_pin_triggers,
-        "outputs": outputs,
-        "mod_guidance": {
-            "enabled": bool(use_anima_mod_guidance),
-            "negative_enabled": bool(use_negative_anima_mod_guidance),
-            "quality_tags": quality_tags,
-            "negative_prompt": negative_quality_tags,
-        },
-        "anima_mod_guidance": {
-            "use_positive": bool(use_anima_mod_guidance),
-            "use_negative": bool(use_negative_anima_mod_guidance),
-            "quality_tags": quality_tags,
-            "negative_prompt": negative_quality_tags,
-        },
-        "artist": {
-            "source": "advanced_artist_field",
-            "handling": "separate" if artist_mix_enabled else "inline",
-            "conditioning_mode": prompt_data_artist_mix_mode if artist_mix_enabled else "none",
-            "include_in_positive": not artist_mix_enabled,
-            "text": positive_artist_inline_prompt,
-            "weighted_text": positive_artist_prompt,
-            "tags": _artist_tags_from_prompt(positive_artist_prompt),
-            "positive_prompt": positive_artist_inline_prompt,
-            "negative_prompt": negative_artist_inline_prompt,
-            "positive_prompt_without_artist": positive_without_artist,
-            "negative_prompt_without_artist": negative_without_artist,
-            "positive_count_hint": len(_parse_artist_mix_items(positive_artist_prompt)),
-            "negative_count_hint": len(_parse_artist_mix_items(negative_artist_prompt)),
-        },
-        "artist_mix": {
-            "enabled": artist_mix_enabled,
-            "mode": prompt_data_artist_mix_mode,
-            "base_source": "positive_without_artist_section",
-            "base_prompt": positive_without_artist,
-            "start_percent": _bounded_artist_mix_float(
-                artist_mix_start_percent,
-                ARTIST_MIX_DEFAULT_START_PERCENT,
-                0.0,
-                1.0,
-            ),
-            "strength_scale": _bounded_artist_mix_float(
-                artist_mix_strength_scale,
-                ARTIST_MIX_DEFAULT_STRENGTH_SCALE,
-                0.0,
-                5.0,
-            ),
-            "style_gain": _bounded_artist_mix_float(
-                artist_mix_style_gain,
-                ARTIST_MIX_DEFAULT_STYLE_GAIN,
-                0.0,
-                3.0,
-            ),
-            "rms_scale_cap": _bounded_artist_mix_float(
-                artist_mix_rms_scale_cap,
-                ARTIST_MIX_DEFAULT_RMS_SCALE_CAP,
-                1.0,
-                5.0,
-            ),
-            "exact_top_k": _bounded_artist_mix_int(
-                artist_mix_exact_top_k,
-                ARTIST_MIX_DEFAULT_EXACT_TOP_K,
-                0,
-                64,
-            ),
-            "cluster_count": _bounded_artist_mix_int(
-                artist_mix_cluster_count,
-                ARTIST_MIX_DEFAULT_CLUSTER_COUNT,
-                1,
-                32,
-            ),
-            "dominant_isolation": _as_bool(
-                artist_mix_dominant_isolation,
-                ARTIST_MIX_DEFAULT_DOMINANT_ISOLATION,
-            ),
-            "dominant_threshold": _bounded_artist_mix_float(
-                artist_mix_dominant_threshold,
-                ARTIST_MIX_DEFAULT_DOMINANT_THRESHOLD,
-                0.0,
-                1.0,
-            ),
-            "artist_prompt": positive_artist_prompt,
-            "artist_count_hint": len(_parse_artist_mix_items(positive_artist_prompt)),
-        },
-        "resolution": {
-            "width": int(width),
-            "height": int(height),
-            "bucket": str(resolution_bucket or DEFAULT_ADVANCED_RESOLUTION_BUCKET),
-            "size": str(resolution_size or ""),
-            "custom_width": _as_int(resolution_custom_width, int(width)),
-            "custom_height": _as_int(resolution_custom_height, int(height)),
-        },
-        "naia": {
-            "use_naia": _as_bool(parameters.get("use_naia"), False),
-            "consume_on_queue": _as_bool(parameters.get("consume_naia_on_queue"), True),
-            "resolution_bucket": str(parameters.get("resolution_bucket") or ""),
-        },
-        "fields": _advanced_prompt_data_fields(effective_fields),
-        "saved_fields": _advanced_prompt_data_fields(saved_fields),
-        "field_inputs": dict(field_inputs),
-        "wildcard": {
-            "mode": str(wildcard_mode or PROMPT_STUDIO_WILDCARD_MODE_LABELS[0]),
-            "seed": normalize_seed(wildcard_seed),
-            "seed_after_generate": str(wildcard_seed_after_generate or SEED_CONTROL_FIXED),
-            "next_seed": wildcard_updates.get("wildcard_seed"),
-            "used_keys": list(wildcard_updates.get("wildcard_used_keys") or []),
-            "missing_keys": list(wildcard_updates.get("wildcard_missing_keys") or []),
-        },
-        "compatibility": {
-            "return_names": list(EasyUseAnimaPromptStudioAdvanced.RETURN_NAMES),
-            "return_types": list(EasyUseAnimaPromptStudioAdvanced.RETURN_TYPES),
-        },
-    }
-
-
 def _get_workflow_node(extra_pnginfo, node_id: str):
     pnginfo = _single_value(extra_pnginfo)
     if not isinstance(pnginfo, dict):
@@ -5171,16 +4529,6 @@ def _get_workflow_node(extra_pnginfo, node_id: str):
             if isinstance(subgraph, dict) and isinstance(subgraph.get("nodes"), list):
                 nodes_list = subgraph["nodes"]
     return found
-
-
-def _normalize_prompt_studio_wildcard_seed_control(value, wildcard_mode=None) -> str:
-    loaded_mode = str(wildcard_mode or "").strip().lower()
-    if loaded_mode in PROMPT_STUDIO_LEGACY_FIXED_WILDCARD_MODES:
-        return SEED_CONTROL_FIXED
-    return PROMPT_STUDIO_WILDCARD_SEED_CONTROL_ALIASES.get(
-        str(value or "").strip().lower(),
-        SEED_CONTROL_FIXED,
-    )
 
 
 def _consume_reserved_wildcard_next_seed(
@@ -5271,6 +4619,13 @@ _bind_regional_node_runtime(
     resolve_helper=lambda name: globals()[name],
     flexible_optional_input_type=_FlexibleOptionalInputType,
 )
+_bind_advanced_runtime(
+    resolve_helper=lambda name: globals()[name],
+)
+_bind_prompt_advanced_node_runtime(
+    resolve_helper=lambda name: globals()[name],
+    flexible_optional_input_type=_FlexibleOptionalInputType,
+)
 _bind_conditioning_runtime(
     resolve_helper=lambda name: globals()[name],
     resolve_logger=lambda: logger,
@@ -5317,767 +4672,6 @@ _bind_lora_node_runtime(
     flexible_optional_input_type=_FlexibleOptionalInputType,
     any_type=_ANY_TYPE,
 )
-
-
-class EasyUseAnimaPromptStudioAdvanced:
-    """Dynamic positive/negative Prompt Studio with serialized field blocks."""
-
-    DESCRIPTION = (
-        "Advanced Prompt Studio with reorderable positive and negative fields, NAIA fill support, "
-        "trigger input handling, Mod Guidance routing, metadata outputs, and latent resolution output."
-    )
-    OUTPUT_TOOLTIPS = (
-        "Final positive prompt assembled from enabled positive fields.",
-        "Final negative prompt assembled from enabled negative fields.",
-        "Positive quality fields routed to Anima Mod Guidance.",
-        "Negative quality fields routed to Anima Mod Guidance.",
-        "Boolean flag passed through for Anima Mod Guidance workflow control.",
-        "Boolean flag passed through for negative Anima Mod Guidance workflow control.",
-        "Positive metadata prompt with metadata filters applied.",
-        "Negative metadata prompt with metadata filters applied.",
-        "Selected latent width.",
-        "Selected latent height.",
-    )
-
-    @classmethod
-    def INPUT_TYPES(cls):
-        return {
-            "required": {
-                "use_naia": ("BOOLEAN", {
-                    "default": False,
-                    "tooltip": (
-                        "Internal request flag. The front-end exposes this as the NAIA Prompt field's "
-                        "'Fill from NAIA' button, which fills that field with a fresh NAIA random prompt."
-                    ),
-                }),
-                "consume_naia_on_queue": ("BOOLEAN", {
-                    "default": True,
-                    "tooltip": (
-                        "Internal one-shot mode. Successful NAIA fills are saved with the request flag off "
-                        "so loaded workflows reuse the stored prompt."
-                    ),
-                }),
-                "use_anima_mod_guidance": ("BOOLEAN", {
-                    "default": False,
-                    "tooltip": (
-                        "true: positive output excludes quality fields and sends them "
-                        "through the mod guidance quality output."
-                    ),
-                }),
-                "resolution_bucket": ("STRING", {
-                    "default": DEFAULT_ADVANCED_RESOLUTION_BUCKET,
-                    "tooltip": "Internal selected latent resolution bucket.",
-                }),
-                "resolution_size": ("STRING", {
-                    "default": DEFAULT_ADVANCED_RESOLUTION_SIZE,
-                    "tooltip": "Internal selected latent resolution, formatted as width * height (ratio).",
-                }),
-                "resolution_custom_width": ("INT", {
-                    "default": 1024,
-                    "min": 32,
-                    "max": 8192,
-                    "step": 32,
-                    "tooltip": "Internal custom latent width. Values are snapped to multiples of 32.",
-                }),
-                "resolution_custom_height": ("INT", {
-                    "default": 1024,
-                    "min": 32,
-                    "max": 8192,
-                    "step": 32,
-                    "tooltip": "Internal custom latent height. Values are snapped to multiples of 32.",
-                }),
-                "pin_trigger_tags_to_front": ("BOOLEAN", {
-                    "default": False,
-                    "tooltip": "Legacy internal flag. Trigger field Pin buttons control trigger placement.",
-                }),
-                "advanced_fields": ("STRING", {
-                    "multiline": True,
-                    "default": _advanced_fields_json(),
-                    "tooltip": "Internal JSON payload for Advanced Prompt Studio fields.",
-                }),
-                "use_negative_anima_mod_guidance": ("BOOLEAN", {
-                    "default": False,
-                    "tooltip": (
-                        "true: negative output excludes negative quality fields and sends them "
-                        "through the negative Mod Guidance output."
-                    ),
-                }),
-                "wildcard_mode": (PROMPT_STUDIO_WILDCARD_MODE_LABELS, {
-                    "default": PROMPT_STUDIO_WILDCARD_MODE_LABELS[0],
-                    "tooltip": "General expands deterministically; Sequential uses seed % option_count. Seed control independently decides the next seed.",
-                }),
-                "wildcard_seed": ("INT", {
-                    "default": 0,
-                    "min": 0,
-                    "max": MAX_SEED,
-                    "tooltip": (
-                        "Wildcard seed used by Advanced Prompt Studio fields. "
-                        f"{WILDCARD_SEED_RANGE_NOTE}"
-                    ),
-                }),
-                "wildcard_seed_after_generate": (SEED_CONTROL_MODES, {
-                    "default": SEED_CONTROL_FIXED,
-                    "tooltip": "After an accepted queue: keep the seed fixed, randomize it, or increment it by one.",
-                }),
-            },
-            "optional": _FlexibleOptionalInputType("STRING"),
-            "hidden": {
-                "workflow_prompt": "PROMPT",
-                "extra_pnginfo": "EXTRA_PNGINFO",
-                "unique_id": "UNIQUE_ID",
-            },
-        }
-
-    RETURN_TYPES = (
-        "STRING",
-        "STRING",
-        "STRING",
-        "STRING",
-        "BOOLEAN",
-        "BOOLEAN",
-        "STRING",
-        "STRING",
-        "INT",
-        "INT",
-    )
-    RETURN_NAMES = (
-        "positive_prompt",
-        "negative_prompt",
-        "anima_mod_guidance_quality_tags",
-        "anima_mod_guidance_negative_prompt",
-        "use_anima_mod_guidance",
-        "use_negative_anima_mod_guidance",
-        "metadata_prompt",
-        "metadata_negative_prompt",
-        "width",
-        "height",
-    )
-    FUNCTION = "build"
-    CATEGORY = "EasyUse Anima/Prompt"
-
-    @classmethod
-    def _widget_input_names(cls) -> list[str]:
-        return list(cls.INPUT_TYPES()["required"].keys())
-
-    @classmethod
-    def IS_CHANGED(
-        cls,
-        use_naia: bool = False,
-        consume_naia_on_queue: bool = True,
-        use_anima_mod_guidance: bool = False,
-        pin_trigger_tags_to_front: bool = False,
-        use_negative_anima_mod_guidance: bool = False,
-        advanced_fields: str = "",
-        resolution_bucket: str = DEFAULT_ADVANCED_RESOLUTION_BUCKET,
-        resolution_size: str = DEFAULT_ADVANCED_RESOLUTION_SIZE,
-        resolution_custom_width: int = 1024,
-        resolution_custom_height: int = 1024,
-        wildcard_mode: str = PROMPT_STUDIO_WILDCARD_MODE_LABELS[0],
-        wildcard_seed: int = 0,
-        wildcard_seed_after_generate: str = SEED_CONTROL_FIXED,
-        **kwargs,
-    ):
-        fields = _normalize_advanced_fields(advanced_fields)
-        if _as_bool(use_naia, False) and (
-            _advanced_has_enabled_naia(fields)
-            or _advanced_uses_naia_resolution(resolution_bucket)
-        ):
-            return float("nan")
-        effective_fields = _apply_advanced_field_inputs(fields, kwargs)
-        wildcard_mode_key = normalize_prompt_studio_wildcard_mode(wildcard_mode)
-        wildcard_active = True
-        wildcard_seed_control = _normalize_prompt_studio_wildcard_seed_control(
-            wildcard_seed_after_generate,
-            wildcard_mode,
-        )
-        return _stable_change_key({
-            "mode": "prompt_studio_advanced",
-            "metadata_filter_words": resolve_metadata_filter_words(),
-            "prompt_translation": _prompt_translation_change_key(),
-            "wildcard_sources": wildcard_sources_signature() if wildcard_active else {},
-            "wildcard_mode": wildcard_mode_key,
-            "wildcard_seed": normalize_seed(wildcard_seed),
-            "wildcard_seed_after_generate": wildcard_seed_control,
-            "use_anima_mod_guidance": _as_bool(use_anima_mod_guidance, False),
-            "use_negative_anima_mod_guidance": _as_bool(use_negative_anima_mod_guidance, False),
-            "resolution": _advanced_resolution_from_selection(
-                resolution_bucket,
-                resolution_size,
-                resolution_custom_width,
-                resolution_custom_height,
-            ),
-            "pin_trigger_tags_to_front": _as_bool(pin_trigger_tags_to_front, False),
-            "advanced_fields": _advanced_fields_json(effective_fields),
-        })
-
-    @classmethod
-    def _update_metadata_fields(
-        cls,
-        workflow_prompt,
-        extra_pnginfo,
-        unique_id,
-        advanced_fields: str,
-        use_naia: bool,
-        extra_updates: dict[str, Any] | None = None,
-    ) -> None:
-        node_id = _single_value(unique_id)
-        if node_id is None:
-            return
-        node_id = str(node_id)
-        updates = {
-            "use_naia": _as_bool(use_naia, False),
-            "advanced_fields": advanced_fields,
-        }
-        if extra_updates:
-            updates.update(extra_updates)
-
-        if isinstance(workflow_prompt, dict):
-            prompt_node = workflow_prompt.get(node_id)
-            if isinstance(prompt_node, dict):
-                inputs = prompt_node.setdefault("inputs", {})
-                for name, value in updates.items():
-                    inputs[name] = value
-
-        workflow_node = _get_workflow_node(extra_pnginfo, node_id)
-        if workflow_node is None:
-            return
-
-        if "advanced_fields" in updates:
-            properties = workflow_node.setdefault("properties", {})
-            if not isinstance(properties, dict):
-                properties = {}
-                workflow_node["properties"] = properties
-            properties[ADVANCED_FIELDS_WORKFLOW_PROPERTY] = updates["advanced_fields"]
-
-        input_names = cls._widget_input_names()
-        widgets_values = workflow_node.setdefault("widgets_values", [])
-        for name, value in updates.items():
-            if name not in input_names:
-                continue
-            index = input_names.index(name)
-            while len(widgets_values) <= index:
-                widgets_values.append(None)
-            widgets_values[index] = value
-
-    @staticmethod
-    def _ui(
-        advanced_fields: str,
-        use_naia: bool,
-        field_inputs: dict | None = None,
-        extra_payload: dict[str, Any] | None = None,
-    ):
-        payload = {
-            "prompt_studio_advanced": [{
-                "advanced_fields": advanced_fields,
-                "use_naia": _as_bool(use_naia, False),
-                "field_inputs": field_inputs or {},
-            }]
-        }
-        if extra_payload:
-            payload["prompt_studio_advanced"][0].update(extra_payload)
-        return payload
-
-    def build(
-        self,
-        use_naia: bool,
-        consume_naia_on_queue: bool,
-        use_anima_mod_guidance: bool,
-        pin_trigger_tags_to_front: bool,
-        advanced_fields: str,
-        use_negative_anima_mod_guidance: bool = False,
-        wildcard_mode: str = PROMPT_STUDIO_WILDCARD_MODE_LABELS[0],
-        wildcard_seed: int = 0,
-        wildcard_seed_after_generate: str = SEED_CONTROL_FIXED,
-        resolution_bucket: str = DEFAULT_ADVANCED_RESOLUTION_BUCKET,
-        resolution_size: str = DEFAULT_ADVANCED_RESOLUTION_SIZE,
-        resolution_custom_width: int = 1024,
-        resolution_custom_height: int = 1024,
-        workflow_prompt=None,
-        extra_pnginfo=None,
-        unique_id=None,
-        **field_inputs,
-    ):
-        fields = _normalize_advanced_fields(advanced_fields)
-        saved_fields = _clone_advanced_fields(fields)
-        effective_field_inputs = _advanced_field_input_values(field_inputs)
-        requested_use_naia = _as_bool(use_naia, False)
-        enabled_naia_panes = _advanced_enabled_naia_panes(fields)
-        use_naia_resolution = _advanced_uses_naia_resolution(resolution_bucket)
-        live_use_naia = requested_use_naia and (bool(enabled_naia_panes) or use_naia_resolution)
-        metadata_use_naia = live_use_naia
-        metadata_updates: dict[str, Any] = {}
-        ui_updates: dict[str, Any] = {}
-        wildcard_mode_key = normalize_prompt_studio_wildcard_mode(wildcard_mode)
-        wildcard_mode_label = (
-            PROMPT_STUDIO_WILDCARD_MODE_LABELS[1]
-            if wildcard_mode_key == WILDCARD_MODE_SEQUENTIAL
-            else PROMPT_STUDIO_WILDCARD_MODE_LABELS[0]
-        )
-        effective_fields = _apply_advanced_field_inputs(fields, effective_field_inputs)
-        wildcard_seed_value = normalize_seed(wildcard_seed)
-        wildcard_effective_seed_control = _normalize_prompt_studio_wildcard_seed_control(
-            wildcard_seed_after_generate,
-            wildcard_mode,
-        )
-        reserved_next_wildcard_seed = _consume_reserved_wildcard_next_seed(
-            field_inputs,
-            workflow_prompt,
-            unique_id,
-            wildcard_seed_value,
-            wildcard_mode_key,
-            wildcard_effective_seed_control,
-        )
-        width, height = _advanced_resolution_from_selection(
-            resolution_bucket,
-            resolution_size,
-            resolution_custom_width,
-            resolution_custom_height,
-        )
-
-        if live_use_naia:
-            naia_settings = resolve_naia_settings()
-            body = EasyUseAnimaNAIARandomPrompt._make_request_body(
-                _as_bool(naia_settings["use_naia_settings"], True),
-                naia_settings["pre_prompt"],
-                naia_settings["post_prompt"],
-                naia_settings["auto_hide"],
-                naia_settings["preprocessing"],
-            )
-            resp = _post_random(
-                naia_settings["host"],
-                naia_settings["port"],
-                body,
-                allow_remote_api=bool(naia_settings.get("allow_remote_api", False)),
-            )
-            naia_prompt, naia_negative, naia_width, naia_height = _parse_random_response(resp)
-            if "positive" in enabled_naia_panes:
-                saved_fields = _set_naia_field_text(saved_fields, "positive", naia_prompt)
-                effective_fields = _set_naia_field_text(effective_fields, "positive", naia_prompt)
-            if "negative" in enabled_naia_panes:
-                saved_fields = _set_naia_field_text(saved_fields, "negative", naia_negative)
-                effective_fields = _set_naia_field_text(effective_fields, "negative", naia_negative)
-            if use_naia_resolution:
-                width, height = _resolve_naia_resolution(naia_width, naia_height, naia_settings)
-                resolution_label = _resolution_label(width, height)
-                metadata_updates.update({
-                    "resolution_bucket": CUSTOM_ADVANCED_RESOLUTION_BUCKET,
-                    "resolution_size": resolution_label,
-                    "resolution_custom_width": width,
-                    "resolution_custom_height": height,
-                })
-                ui_updates.update({
-                    "resolution_bucket": NAIA_ADVANCED_RESOLUTION_BUCKET,
-                    "resolution_size": resolution_label,
-                    "resolution_custom_width": width,
-                    "resolution_custom_height": height,
-                })
-            metadata_use_naia = False
-
-        ui_fields = _clone_advanced_fields(saved_fields)
-        effective_fields, effective_wildcard = _expand_advanced_wildcard_fields(
-            effective_fields,
-            wildcard_seed_value,
-            wildcard_mode_key,
-        )
-        effective_fields = _translate_prompt_fields(effective_fields)
-        next_wildcard_seed = (
-            reserved_next_wildcard_seed
-            if reserved_next_wildcard_seed is not None
-            else next_seed(wildcard_seed_value, wildcard_effective_seed_control)
-        )
-        ui_updates.update({
-            "wildcard_mode": wildcard_mode_label,
-            "wildcard_seed": next_wildcard_seed,
-            "wildcard_seed_after_generate": wildcard_effective_seed_control,
-            "wildcard_used_keys": list(effective_wildcard["used_keys"]),
-            "wildcard_missing_keys": list(effective_wildcard["missing_keys"]),
-        })
-        metadata_updates.update({
-            "wildcard_mode": wildcard_mode_label,
-            "wildcard_seed": wildcard_seed_value,
-            "wildcard_seed_after_generate": SEED_CONTROL_FIXED,
-        })
-
-        fields_json = _advanced_fields_json(saved_fields)
-        ui_fields_json = _advanced_fields_json(ui_fields)
-        if live_use_naia or metadata_updates:
-            self._update_metadata_fields(
-                workflow_prompt,
-                extra_pnginfo,
-                unique_id,
-                fields_json,
-                metadata_use_naia,
-                metadata_updates,
-            )
-
-        result = _build_advanced_prompts(
-            effective_fields,
-            use_anima_mod_guidance,
-            use_negative_anima_mod_guidance,
-            pin_trigger_tags_to_front,
-        )
-        return {
-            "ui": self._ui(ui_fields_json, requested_use_naia, effective_field_inputs, ui_updates),
-            "result": (*result, width, height),
-        }
-
-
-class EasyUseAnimaPromptStudioAdvancedV2(EasyUseAnimaPromptStudioAdvanced):
-    """Advanced Prompt Studio v2 with structured prompt data output."""
-
-    DESCRIPTION = (
-        "Advanced Prompt Studio v2. It outputs a single EASYUSE_ANIMA_PROMPT_DATA "
-        "dict socket for downstream nodes."
-    )
-    OUTPUT_TOOLTIPS = ("Structured prompt data dict for downstream EasyUse Anima nodes.",)
-    RETURN_TYPES = (PROMPT_DATA_TYPE,)
-    RETURN_NAMES = (PROMPT_DATA_TYPE,)
-
-    @classmethod
-    def INPUT_TYPES(cls):
-        base = EasyUseAnimaPromptStudioAdvanced.INPUT_TYPES()
-        required = dict(base["required"])
-        required.update({
-            "artist_mix_mode": (list(ARTIST_MIX_STUDIO_MODES), {
-                "default": ARTIST_MIX_MODE_OFF,
-                "tooltip": _artist_mix_mode_tooltip(),
-            }),
-            "artist_mix_start_percent": ("FLOAT", {
-                "default": ARTIST_MIX_DEFAULT_START_PERCENT,
-                "min": 0.0,
-                "max": 1.0,
-                "step": 0.01,
-                "tooltip": "Start percent used by late/scheduled artist mix modes.",
-            }),
-            "artist_mix_strength_scale": ("FLOAT", {
-                "default": ARTIST_MIX_DEFAULT_STRENGTH_SCALE,
-                "min": 0.0,
-                "max": 5.0,
-                "step": 0.01,
-                "tooltip": "Strength multiplier used by exact artist mix modes.",
-            }),
-            "artist_mix_style_gain": ("FLOAT", {
-                "default": ARTIST_MIX_DEFAULT_STYLE_GAIN,
-                "min": 0.0,
-                "max": 3.0,
-                "step": 0.01,
-                "tooltip": "Style delta gain used by delta_rms, hybrid tail, and clustered compressed branches.",
-            }),
-            "artist_mix_rms_scale_cap": ("FLOAT", {
-                "default": ARTIST_MIX_DEFAULT_RMS_SCALE_CAP,
-                "min": 1.0,
-                "max": 5.0,
-                "step": 0.01,
-                "tooltip": "Maximum RMS energy restore scale for delta_rms compressed artist branches.",
-            }),
-            "artist_mix_exact_top_k": ("INT", {
-                "default": ARTIST_MIX_DEFAULT_EXACT_TOP_K,
-                "min": 0,
-                "max": 64,
-                "tooltip": "Hybrid mode keeps this many strongest artists as exact positive branches.",
-            }),
-            "artist_mix_cluster_count": ("INT", {
-                "default": ARTIST_MIX_DEFAULT_CLUSTER_COUNT,
-                "min": 1,
-                "max": 32,
-                "tooltip": "Clustered mode compresses non-dominant artists into this many positive branches.",
-            }),
-            "artist_mix_dominant_isolation": ("BOOLEAN", {
-                "default": ARTIST_MIX_DEFAULT_DOMINANT_ISOLATION,
-                "tooltip": "Clustered mode keeps artists above the dominant threshold as exact branches.",
-            }),
-            "artist_mix_dominant_threshold": ("FLOAT", {
-                "default": ARTIST_MIX_DEFAULT_DOMINANT_THRESHOLD,
-                "min": 0.0,
-                "max": 1.0,
-                "step": 0.01,
-                "tooltip": "Clustered dominant isolation threshold based on normalized artist weight.",
-            }),
-        })
-        return {
-            **base,
-            "required": required,
-        }
-
-    @classmethod
-    def IS_CHANGED(
-        cls,
-        use_naia: bool = False,
-        consume_naia_on_queue: bool = True,
-        use_anima_mod_guidance: bool = False,
-        pin_trigger_tags_to_front: bool = False,
-        use_negative_anima_mod_guidance: bool = False,
-        advanced_fields: str = "",
-        resolution_bucket: str = DEFAULT_ADVANCED_RESOLUTION_BUCKET,
-        resolution_size: str = DEFAULT_ADVANCED_RESOLUTION_SIZE,
-        resolution_custom_width: int = 1024,
-        resolution_custom_height: int = 1024,
-        wildcard_mode: str = PROMPT_STUDIO_WILDCARD_MODE_LABELS[0],
-        wildcard_seed: int = 0,
-        wildcard_seed_after_generate: str = SEED_CONTROL_FIXED,
-        artist_mix_mode: str = ARTIST_MIX_MODE_OFF,
-        artist_mix_start_percent: float = ARTIST_MIX_DEFAULT_START_PERCENT,
-        artist_mix_strength_scale: float = ARTIST_MIX_DEFAULT_STRENGTH_SCALE,
-        artist_mix_style_gain: float = ARTIST_MIX_DEFAULT_STYLE_GAIN,
-        artist_mix_rms_scale_cap: float = ARTIST_MIX_DEFAULT_RMS_SCALE_CAP,
-        artist_mix_exact_top_k: int = ARTIST_MIX_DEFAULT_EXACT_TOP_K,
-        artist_mix_cluster_count: int = ARTIST_MIX_DEFAULT_CLUSTER_COUNT,
-        artist_mix_dominant_isolation: bool = ARTIST_MIX_DEFAULT_DOMINANT_ISOLATION,
-        artist_mix_dominant_threshold: float = ARTIST_MIX_DEFAULT_DOMINANT_THRESHOLD,
-        **field_inputs,
-    ):
-        base_key = EasyUseAnimaPromptStudioAdvanced.IS_CHANGED(
-            use_naia=use_naia,
-            consume_naia_on_queue=consume_naia_on_queue,
-            use_anima_mod_guidance=use_anima_mod_guidance,
-            pin_trigger_tags_to_front=pin_trigger_tags_to_front,
-            use_negative_anima_mod_guidance=use_negative_anima_mod_guidance,
-            advanced_fields=advanced_fields,
-            resolution_bucket=resolution_bucket,
-            resolution_size=resolution_size,
-            resolution_custom_width=resolution_custom_width,
-            resolution_custom_height=resolution_custom_height,
-            wildcard_mode=wildcard_mode,
-            wildcard_seed=wildcard_seed,
-            wildcard_seed_after_generate=wildcard_seed_after_generate,
-            **field_inputs,
-        )
-        if base_key != base_key:
-            return base_key
-        return _stable_change_key({
-            "base": base_key,
-            "artist_mix_mode": _normalize_artist_mix_mode(artist_mix_mode, ARTIST_MIX_MODE_OFF),
-            "artist_mix_start_percent": _bounded_artist_mix_float(
-                artist_mix_start_percent,
-                ARTIST_MIX_DEFAULT_START_PERCENT,
-                0.0,
-                1.0,
-            ),
-            "artist_mix_strength_scale": _bounded_artist_mix_float(
-                artist_mix_strength_scale,
-                ARTIST_MIX_DEFAULT_STRENGTH_SCALE,
-                0.0,
-                5.0,
-            ),
-            "artist_mix_style_gain": _bounded_artist_mix_float(
-                artist_mix_style_gain,
-                ARTIST_MIX_DEFAULT_STYLE_GAIN,
-                0.0,
-                3.0,
-            ),
-            "artist_mix_rms_scale_cap": _bounded_artist_mix_float(
-                artist_mix_rms_scale_cap,
-                ARTIST_MIX_DEFAULT_RMS_SCALE_CAP,
-                1.0,
-                5.0,
-            ),
-            "artist_mix_exact_top_k": _bounded_artist_mix_int(
-                artist_mix_exact_top_k,
-                ARTIST_MIX_DEFAULT_EXACT_TOP_K,
-                0,
-                64,
-            ),
-            "artist_mix_cluster_count": _bounded_artist_mix_int(
-                artist_mix_cluster_count,
-                ARTIST_MIX_DEFAULT_CLUSTER_COUNT,
-                1,
-                32,
-            ),
-            "artist_mix_dominant_isolation": _as_bool(
-                artist_mix_dominant_isolation,
-                ARTIST_MIX_DEFAULT_DOMINANT_ISOLATION,
-            ),
-            "artist_mix_dominant_threshold": _bounded_artist_mix_float(
-                artist_mix_dominant_threshold,
-                ARTIST_MIX_DEFAULT_DOMINANT_THRESHOLD,
-                0.0,
-                1.0,
-            ),
-        })
-
-    def build(
-        self,
-        use_naia: bool,
-        consume_naia_on_queue: bool,
-        use_anima_mod_guidance: bool,
-        pin_trigger_tags_to_front: bool,
-        advanced_fields: str,
-        use_negative_anima_mod_guidance: bool = False,
-        wildcard_mode: str = PROMPT_STUDIO_WILDCARD_MODE_LABELS[0],
-        wildcard_seed: int = 0,
-        wildcard_seed_after_generate: str = SEED_CONTROL_FIXED,
-        resolution_bucket: str = DEFAULT_ADVANCED_RESOLUTION_BUCKET,
-        resolution_size: str = DEFAULT_ADVANCED_RESOLUTION_SIZE,
-        resolution_custom_width: int = 1024,
-        resolution_custom_height: int = 1024,
-        artist_mix_mode: str = ARTIST_MIX_MODE_OFF,
-        artist_mix_start_percent: float = ARTIST_MIX_DEFAULT_START_PERCENT,
-        artist_mix_strength_scale: float = ARTIST_MIX_DEFAULT_STRENGTH_SCALE,
-        artist_mix_style_gain: float = ARTIST_MIX_DEFAULT_STYLE_GAIN,
-        artist_mix_rms_scale_cap: float = ARTIST_MIX_DEFAULT_RMS_SCALE_CAP,
-        artist_mix_exact_top_k: int = ARTIST_MIX_DEFAULT_EXACT_TOP_K,
-        artist_mix_cluster_count: int = ARTIST_MIX_DEFAULT_CLUSTER_COUNT,
-        artist_mix_dominant_isolation: bool = ARTIST_MIX_DEFAULT_DOMINANT_ISOLATION,
-        artist_mix_dominant_threshold: float = ARTIST_MIX_DEFAULT_DOMINANT_THRESHOLD,
-        workflow_prompt=None,
-        extra_pnginfo=None,
-        unique_id=None,
-        **field_inputs,
-    ):
-        base = super().build(
-            use_naia,
-            consume_naia_on_queue,
-            use_anima_mod_guidance,
-            pin_trigger_tags_to_front,
-            advanced_fields,
-            use_negative_anima_mod_guidance=use_negative_anima_mod_guidance,
-            wildcard_mode=wildcard_mode,
-            wildcard_seed=wildcard_seed,
-            wildcard_seed_after_generate=wildcard_seed_after_generate,
-            resolution_bucket=resolution_bucket,
-            resolution_size=resolution_size,
-            resolution_custom_width=resolution_custom_width,
-            resolution_custom_height=resolution_custom_height,
-            workflow_prompt=workflow_prompt,
-            extra_pnginfo=extra_pnginfo,
-            unique_id=unique_id,
-            **field_inputs,
-        )
-        compat_result = tuple(base.get("result") or ())
-        ui_payloads = base.get("ui", {}).get("prompt_studio_advanced", [])
-        ui_payload = ui_payloads[0] if ui_payloads and isinstance(ui_payloads[0], dict) else {}
-        if isinstance(ui_payload, dict):
-            ui_payload.update({
-                "artist_mix_mode": _normalize_artist_mix_mode(artist_mix_mode, ARTIST_MIX_MODE_OFF),
-                "artist_mix_start_percent": _bounded_artist_mix_float(
-                    artist_mix_start_percent,
-                    ARTIST_MIX_DEFAULT_START_PERCENT,
-                    0.0,
-                    1.0,
-                ),
-                "artist_mix_strength_scale": _bounded_artist_mix_float(
-                    artist_mix_strength_scale,
-                    ARTIST_MIX_DEFAULT_STRENGTH_SCALE,
-                    0.0,
-                    5.0,
-                ),
-                "artist_mix_style_gain": _bounded_artist_mix_float(
-                    artist_mix_style_gain,
-                    ARTIST_MIX_DEFAULT_STYLE_GAIN,
-                    0.0,
-                    3.0,
-                ),
-                "artist_mix_rms_scale_cap": _bounded_artist_mix_float(
-                    artist_mix_rms_scale_cap,
-                    ARTIST_MIX_DEFAULT_RMS_SCALE_CAP,
-                    1.0,
-                    5.0,
-                ),
-                "artist_mix_exact_top_k": _bounded_artist_mix_int(
-                    artist_mix_exact_top_k,
-                    ARTIST_MIX_DEFAULT_EXACT_TOP_K,
-                    0,
-                    64,
-                ),
-                "artist_mix_cluster_count": _bounded_artist_mix_int(
-                    artist_mix_cluster_count,
-                    ARTIST_MIX_DEFAULT_CLUSTER_COUNT,
-                    1,
-                    32,
-                ),
-                "artist_mix_dominant_isolation": _as_bool(
-                    artist_mix_dominant_isolation,
-                    ARTIST_MIX_DEFAULT_DOMINANT_ISOLATION,
-                ),
-                "artist_mix_dominant_threshold": _bounded_artist_mix_float(
-                    artist_mix_dominant_threshold,
-                    ARTIST_MIX_DEFAULT_DOMINANT_THRESHOLD,
-                    0.0,
-                    1.0,
-                ),
-            })
-        saved_fields = _normalize_advanced_fields(ui_payload.get("advanced_fields", advanced_fields))
-        effective_field_inputs = _advanced_field_input_values(ui_payload.get("field_inputs") or field_inputs)
-        wildcard_mode_key = normalize_prompt_studio_wildcard_mode(wildcard_mode)
-        effective_fields = _apply_advanced_field_inputs(saved_fields, effective_field_inputs)
-        effective_fields, _wildcard = _expand_advanced_wildcard_fields(
-            effective_fields,
-            normalize_seed(wildcard_seed),
-            wildcard_mode_key,
-        )
-        effective_fields = _translate_prompt_fields(effective_fields)
-        prompt_data_parameters = _prompt_data_parameter_snapshot(
-            self.INPUT_TYPES().get("required", {}),
-            {
-                "use_naia": use_naia,
-                "consume_naia_on_queue": consume_naia_on_queue,
-                "use_anima_mod_guidance": use_anima_mod_guidance,
-                "pin_trigger_tags_to_front": pin_trigger_tags_to_front,
-                "advanced_fields": advanced_fields,
-                "use_negative_anima_mod_guidance": use_negative_anima_mod_guidance,
-                "wildcard_mode": (
-                    PROMPT_STUDIO_WILDCARD_MODE_LABELS[1]
-                    if wildcard_mode_key == WILDCARD_MODE_SEQUENTIAL
-                    else PROMPT_STUDIO_WILDCARD_MODE_LABELS[0]
-                ),
-                "wildcard_seed": wildcard_seed,
-                "wildcard_seed_after_generate": _normalize_prompt_studio_wildcard_seed_control(
-                    wildcard_seed_after_generate,
-                    wildcard_mode,
-                ),
-                "resolution_bucket": resolution_bucket,
-                "resolution_size": resolution_size,
-                "resolution_custom_width": resolution_custom_width,
-                "resolution_custom_height": resolution_custom_height,
-                "artist_mix_mode": artist_mix_mode,
-                "artist_mix_start_percent": artist_mix_start_percent,
-                "artist_mix_strength_scale": artist_mix_strength_scale,
-                "artist_mix_style_gain": artist_mix_style_gain,
-                "artist_mix_rms_scale_cap": artist_mix_rms_scale_cap,
-                "artist_mix_exact_top_k": artist_mix_exact_top_k,
-                "artist_mix_cluster_count": artist_mix_cluster_count,
-                "artist_mix_dominant_isolation": artist_mix_dominant_isolation,
-                "artist_mix_dominant_threshold": artist_mix_dominant_threshold,
-                **field_inputs,
-            },
-            ui_payload,
-        )
-        prompt_data_parameters["wildcard_seed"] = normalize_seed(wildcard_seed)
-        prompt_data = _build_advanced_prompt_data(
-            compat_result,
-            effective_fields,
-            saved_fields,
-            effective_field_inputs,
-            str(ui_payload.get("resolution_bucket", resolution_bucket)),
-            str(ui_payload.get("resolution_size", resolution_size)),
-            _as_int(ui_payload.get("resolution_custom_width", resolution_custom_width), resolution_custom_width),
-            _as_int(ui_payload.get("resolution_custom_height", resolution_custom_height), resolution_custom_height),
-            str(ui_payload.get("wildcard_mode", wildcard_mode)),
-            normalize_seed(wildcard_seed),
-            str(ui_payload.get("wildcard_seed_after_generate", wildcard_seed_after_generate)),
-            ui_payload,
-            pin_trigger_tags_to_front,
-            parameters=prompt_data_parameters,
-            artist_mix_mode=artist_mix_mode,
-            artist_mix_start_percent=artist_mix_start_percent,
-            artist_mix_strength_scale=artist_mix_strength_scale,
-            artist_mix_style_gain=artist_mix_style_gain,
-            artist_mix_rms_scale_cap=artist_mix_rms_scale_cap,
-            artist_mix_exact_top_k=artist_mix_exact_top_k,
-            artist_mix_cluster_count=artist_mix_cluster_count,
-            artist_mix_dominant_isolation=artist_mix_dominant_isolation,
-            artist_mix_dominant_threshold=artist_mix_dominant_threshold,
-        )
-        return {
-            **base,
-            "result": (prompt_data,),
-        }
-
-
-
-
-
-
 
 
 def _easy_use_anima_input_signature(value) -> dict[str, Any]:
@@ -6656,257 +5250,6 @@ class EasyUseAnimaAIOGenerator:
         return {
             "ui": ui,
             "result": (image, latent, metadata_json),
-        }
-
-
-class EasyUseAnimaPromptStudioExtend:
-    """Extended Prompt Studio with numbered positive/negative prompt rows."""
-
-    @classmethod
-    def INPUT_TYPES(cls):
-        required = {
-            "fill_naia_prompt": ("BOOLEAN", {
-                "default": False,
-                "tooltip": (
-                    "When enabled, slot 3 is filled from NAIA on each queue. "
-                    "Saved-image workflows are written with this off and the current slot 3 text stored."
-                ),
-            }),
-            "use_anima_mod_guidance": ("BOOLEAN", {
-                "default": False,
-                "tooltip": (
-                    "true: positive output excludes slots 1-2 quality tags and sends them "
-                    "through anima_mod_guidance_quality_tags."
-                ),
-            }),
-            "pin_trigger_tags_to_front": ("BOOLEAN", {
-                "default": False,
-                "tooltip": "true: keep prompt correction trigger-like tags at the front where applicable.",
-            }),
-        }
-        for name, _pane, _field_type, label, default, height in EXTEND_PROMPT_SLOT_SPECS:
-            required[name] = ("STRING", {
-                "multiline": True,
-                "default": default,
-                "tooltip": label,
-                "placeholder": label,
-                "height": height,
-            })
-        required["active_slots"] = ("STRING", {
-            "default": json.dumps(["quality_tags_1", "general_tags_4", "trailing_tags_10"]),
-            "tooltip": "Internal Prompt Studio Extend visible slot state.",
-        })
-        required["use_negative_anima_mod_guidance"] = ("BOOLEAN", {
-            "default": False,
-            "tooltip": (
-                "true: negative output excludes negative quality slots and sends them "
-                "through the negative Mod Guidance output."
-            ),
-        })
-        return {
-            "required": required,
-            "hidden": {
-                "workflow_prompt": "PROMPT",
-                "extra_pnginfo": "EXTRA_PNGINFO",
-                "unique_id": "UNIQUE_ID",
-            },
-        }
-
-    RETURN_TYPES = (
-        "STRING",
-        "STRING",
-        "STRING",
-        "STRING",
-        "BOOLEAN",
-        "BOOLEAN",
-        "STRING",
-        "STRING",
-    )
-    RETURN_NAMES = (
-        "positive_prompt",
-        "negative_prompt",
-        "anima_mod_guidance_quality_tags",
-        "anima_mod_guidance_negative_prompt",
-        "use_anima_mod_guidance",
-        "use_negative_anima_mod_guidance",
-        "metadata_prompt",
-        "metadata_negative_prompt",
-    )
-    FUNCTION = "build"
-    CATEGORY = "EasyUse Anima/Prompt"
-
-    @classmethod
-    def _widget_input_names(cls) -> list[str]:
-        return list(cls.INPUT_TYPES()["required"].keys())
-
-    @classmethod
-    def IS_CHANGED(
-        cls,
-        fill_naia_prompt: bool = False,
-        use_anima_mod_guidance: bool = False,
-        pin_trigger_tags_to_front: bool = False,
-        use_negative_anima_mod_guidance: bool = False,
-        **kwargs,
-    ):
-        if _as_bool(fill_naia_prompt, False):
-            return float("nan")
-        return _stable_change_key({
-            "mode": "prompt_studio_extend",
-            "metadata_filter_words": resolve_metadata_filter_words(),
-            "prompt_translation": _prompt_translation_change_key(),
-            "use_anima_mod_guidance": _as_bool(use_anima_mod_guidance, False),
-            "use_negative_anima_mod_guidance": _as_bool(use_negative_anima_mod_guidance, False),
-            "pin_trigger_tags_to_front": _as_bool(pin_trigger_tags_to_front, False),
-            "active_slots": str(kwargs.get("active_slots", "")),
-            **{
-                name: str(kwargs.get(name, ""))
-                for name, *_rest in EXTEND_PROMPT_SLOT_SPECS
-            },
-        })
-
-    @classmethod
-    def _update_metadata_slots(
-        cls,
-        workflow_prompt,
-        extra_pnginfo,
-        unique_id,
-        updates: dict[str, Any],
-    ) -> None:
-        node_id = _single_value(unique_id)
-        if node_id is None:
-            return
-        node_id = str(node_id)
-
-        if isinstance(workflow_prompt, dict):
-            prompt_node = workflow_prompt.get(node_id)
-            if isinstance(prompt_node, dict):
-                inputs = prompt_node.setdefault("inputs", {})
-                for name, value in updates.items():
-                    inputs[name] = value
-
-        workflow_node = _get_workflow_node(extra_pnginfo, node_id)
-        if workflow_node is None:
-            return
-
-        input_names = cls._widget_input_names()
-        widgets_values = workflow_node.setdefault("widgets_values", [])
-        for name, value in updates.items():
-            if name not in input_names:
-                continue
-            index = input_names.index(name)
-            while len(widgets_values) <= index:
-                widgets_values.append(None)
-            widgets_values[index] = value
-
-    @staticmethod
-    @staticmethod
-    def _active_slot_set(active_slots: Any) -> set[str] | None:
-        if active_slots is None:
-            return None
-        parsed = active_slots
-        if isinstance(active_slots, str):
-            if not active_slots.strip():
-                return None
-            try:
-                parsed = json.loads(active_slots)
-            except json.JSONDecodeError:
-                return None
-        if not isinstance(parsed, list):
-            return None
-        valid_names = {name for name, *_rest in EXTEND_PROMPT_SLOT_SPECS}
-        return {str(name) for name in parsed if str(name) in valid_names}
-
-    @staticmethod
-    def _fields_from_slots(values: dict[str, str], active_slots: Any = None) -> list[dict]:
-        active_slot_names = EasyUseAnimaPromptStudioExtend._active_slot_set(active_slots)
-        fields = []
-        for name, pane, field_type, label, _default, height in EXTEND_PROMPT_SLOT_SPECS:
-            if active_slot_names is not None and name not in active_slot_names:
-                continue
-            text = str(values.get(name, "") or "")
-            fields.append({
-                "id": name,
-                "pane": pane,
-                "type": field_type,
-                "label": label,
-                "text": text,
-                "height": height,
-                "enabled": bool(text.strip()),
-            })
-        return fields
-
-    @staticmethod
-    def _ui(slot_values: dict[str, str], fill_naia_prompt: bool, active_slots: Any = None):
-        return {
-            "prompt_studio_slots": [{
-                **slot_values,
-                "fill_naia_prompt": _as_bool(fill_naia_prompt, False),
-                "active_slots": active_slots,
-            }]
-        }
-
-    def build(
-        self,
-        fill_naia_prompt: bool,
-        use_anima_mod_guidance: bool,
-        pin_trigger_tags_to_front: bool,
-        use_negative_anima_mod_guidance: bool = False,
-        workflow_prompt=None,
-        extra_pnginfo=None,
-        unique_id=None,
-        **slot_values,
-    ):
-        active_slots = slot_values.get("active_slots")
-        values = {
-            name: str(slot_values.get(name, default) or "")
-            for name, _pane, _field_type, _label, default, _height in EXTEND_PROMPT_SLOT_SPECS
-        }
-        live_fill_naia = _as_bool(fill_naia_prompt, False)
-        metadata_fill_naia = live_fill_naia
-
-        if live_fill_naia:
-            naia_settings = resolve_naia_settings()
-            body = EasyUseAnimaNAIARandomPrompt._make_request_body(
-                _as_bool(naia_settings["use_naia_settings"], True),
-                naia_settings["pre_prompt"],
-                naia_settings["post_prompt"],
-                naia_settings["auto_hide"],
-                naia_settings["preprocessing"],
-            )
-            resp = _post_random(
-                naia_settings["host"],
-                naia_settings["port"],
-                body,
-                allow_remote_api=bool(naia_settings.get("allow_remote_api", False)),
-            )
-            naia_prompt, _naia_negative, _naia_width, _naia_height = _parse_random_response(resp)
-            values["naia_prompt_3"] = naia_prompt
-            metadata_fill_naia = False
-
-        if live_fill_naia:
-            self._update_metadata_slots(
-                workflow_prompt,
-                extra_pnginfo,
-                unique_id,
-                {
-                    "fill_naia_prompt": metadata_fill_naia,
-                    "naia_prompt_3": values["naia_prompt_3"],
-                },
-            )
-
-        effective_values = {
-            name: _translate_prompt_text(value)
-            for name, value in values.items()
-        }
-        result = _build_advanced_prompts(
-            self._fields_from_slots(effective_values, active_slots),
-            use_anima_mod_guidance,
-            use_negative_anima_mod_guidance,
-            pin_trigger_tags_to_front,
-        )
-        return {
-            "ui": self._ui(values, live_fill_naia, active_slots),
-            "result": result,
         }
