+++ tests/test_python_backend_analyzer.py
@@ -683,9 +683,9 @@ ignored/
 
         self.assertEqual(analyzer.render_json(report), expected_text)
         self.assertEqual(report["schema_version"], 2)
-        self.assertEqual(report["inventory"]["module_count"], 55)
-        self.assertEqual(len(report["registry"]["shipped_python_modules"]), 55)
-        self.assertEqual(len(report["registry"]["runtime_import_closure"]), 54)
+        self.assertEqual(report["inventory"]["module_count"], 57)
+        self.assertEqual(len(report["registry"]["shipped_python_modules"]), 57)
+        self.assertEqual(len(report["registry"]["runtime_import_closure"]), 56)
         self.assertEqual(report["registry"]["missing_internal_imports"], [])
         self.assertEqual(
             report["registry"]["unreachable_shipped_python_modules"],
@@ -719,6 +719,7 @@ ignored/
                 "easyuse_anima/nodes/image_nodes.py",
                 "easyuse_anima/nodes/lora_nodes.py",
                 "easyuse_anima/nodes/naia_nodes.py",
+                "easyuse_anima/nodes/prompt_advanced_nodes.py",
                 "easyuse_anima/nodes/prompt_data_nodes.py",
                 "easyuse_anima/nodes/prompt_nodes.py",
                 "easyuse_anima/nodes/regional_nodes.py",
@@ -727,6 +728,7 @@ ignored/
                 "easyuse_anima/profiles/contract.py",
                 "easyuse_anima/profiles/mutation.py",
                 "easyuse_anima/prompt/__init__.py",
+                "easyuse_anima/prompt/advanced.py",
                 "easyuse_anima/prompt/correction.py",
                 "easyuse_anima/prompt/artist_mix.py",
                 "easyuse_anima/prompt/conditioning.py",
@@ -756,11 +758,13 @@ ignored/
                 "easyuse_anima/nodes/image_nodes.py",
                 "easyuse_anima/nodes/lora_nodes.py",
                 "easyuse_anima/nodes/naia_nodes.py",
+                "easyuse_anima/nodes/prompt_advanced_nodes.py",
                 "easyuse_anima/nodes/prompt_data_nodes.py",
                 "easyuse_anima/nodes/prompt_nodes.py",
                 "easyuse_anima/nodes/regional_nodes.py",
                 "easyuse_anima/nodes/wildcard_nodes.py",
                 "easyuse_anima/prompt/__init__.py",
+                "easyuse_anima/prompt/advanced.py",
                 "easyuse_anima/prompt/correction.py",
                 "easyuse_anima/prompt/artist_mix.py",
                 "easyuse_anima/prompt/conditioning.py",
