+++ tests/test_python_package_skeleton.py
@@ -24,12 +24,14 @@ PACKAGE_MODULES = (
     "easyuse_anima.lora",
     "easyuse_anima.naia",
     "easyuse_anima.nodes",
+    "easyuse_anima.nodes.prompt_advanced_nodes",
     "easyuse_anima.nodes.prompt_data_nodes",
     "easyuse_anima.nodes.regional_nodes",
     "easyuse_anima.profiles",
     "easyuse_anima.profiles.contract",
     "easyuse_anima.profiles.mutation",
     "easyuse_anima.prompt",
+    "easyuse_anima.prompt.advanced",
     "easyuse_anima.prompt.artist_mix",
     "easyuse_anima.prompt.conditioning",
     "easyuse_anima.prompt.data",
